telegram-llm-aws/requirements.txt
python-telegram-bot==20.7
openai>=1.0
redis
python-dotenv